# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Event Management API
  description: An API developed for our WBD project that serves our project.
  version: 1.0.0
servers:
  - url: http://chomspro.herokuapp.com
    description:  Main (production) Backend server
  - url: http://localhost:3000
    description:  Main (production) Frontend server
...
paths:
  /booking:
    get:
      summary: Lists all the bookings done till now 
      tags: [Bookings]
      responses:
        "200":
          description: The list of all bookings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bookings'
    post:
      summary: Adds a new booking record.
      tags: [Bookings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Bookings'
      responses: 
        '200':
          description: Created
        '400':
          description: Error
  /booking/{bookingID}:
    get:
      summary: Returns information of a booking by ID.
      tags: [Bookings]
      parameters:
        - name: bookingID
          in: path
          required: true
          description: Autogenerated Booking ID by MongoDB
          schema:
            type : string
      responses: 
        '200':
          description: OK
        '400':
          description: Record does not exist
        '500':
          description: Error

    delete:
      summary: Deletes a booking by ID.
      tags: [Bookings]
      parameters:
        - name: bookingID
          in: path
          required: true
          description: Autogenerated Booking ID by MongoDB
          schema:
            type : string
      responses: 
        '200':
          description: Deleted
        '400':
          description: Record does not exist
        '500':
          description: Error
          
    patch:
      summary: Update a booking information
      tags: [Bookings]
      parameters:
        - name: bookingID
          in: path
          required: true
          description: Autogenerated Booking ID by MongoDB
          schema:
            type : string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Bookings'
      responses:
        '200':
          description: Booking Data Updated
          schema:
            $ref: "#/components/schemas/Bookings"
        '400':
          description: Error



# -------------------------------------------------------------------------
# EVENT
# -------------------------------------------------------------------------
  /events:
    get:
      summary: Lists all the events done till now 
      tags: [Events]
      responses:
        "200":
          description: The list of all events.
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Events'
    post:
      summary: Adds a new event record.
      tags: [Events]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Events'
      responses: 
        '200':
          description: Created
        '400':
          description: Error 
  /events/{eventsID}:
    get:
      summary: Returns information of a Event by ID.
      tags: [Events]
      parameters:
        - name: eventsID
          in: path
          required: true
          description: Autogenerated Event ID by MongoDB
          schema:
            type : string
      responses: 
        '200':
          description: OK
        '400':
          description: Record does not exist
        '500':
          description: Error

    delete:
      summary: Deletes a event by ID.
      tags: [Events]
      parameters:
        - name: eventsID
          in: path
          required: true
          description: Autogenerated Events ID by MongoDB
          schema:
            type : string
      responses: 
        '200':
          description: Deleted
        '400':
          description: Record does not exist
        '500':
          description: Error
          
    patch:
      summary: Update a events information
      tags: [Events]
      parameters:
        - name: eventsID
          in: path
          required: true
          description: Autogenerated Events ID by MongoDB
          schema:
            type : string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Events'
      responses:
        '200':
          description: Booking Data Updated
          schema:
            $ref: "#/components/schemas/Events"
        '400':
          description: Error


# -------------------------------------------------------------------------
# FEED
# -------------------------------------------------------------------------
  /feed:
    get:
      summary: Lists all the Feeds done till now 
      tags: [Feeds]
      responses:
        "200":
          description: The list of all feeds.
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Feeds'  
    post:
      summary: Adds a new Feed record.
      tags: [Feeds]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Feeds'
      responses: 
        '200':
          description: Created
        '400':
          description: Error 
  /feed/{feedID}:
    get:
      summary: Returns information of a Feed by ID.
      tags: [Feeds]
      parameters:
        - name: FeedID
          in: path
          required: true
          description: Autogenerated Feed ID by MongoDB
          schema:
            type : string
      responses: 
        '200':
          description: OK
        '400':
          description: Record does not exist
        '500':
          description: Error

    delete:
      summary: Deletes a Feed by ID.
      tags: [Feeds]
      parameters:
        - name: FeedID
          in: path
          required: true
          description: Autogenerated Feed ID by MongoDB
          schema:
            type : string
      responses: 
        '200':
          description: Deleted
        '400':
          description: Record does not exist
        '500':
          description: Error
          
    patch:
      summary: Update a feed information
      tags: [Feeds]
      parameters:
        - name: FeedID
          in: path
          required: true
          description: Autogenerated Feeds ID by MongoDB
          schema:
            type : string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Feeds'
      responses:
        '200':
          description: Feed Data Updated
          schema:
            $ref: "#/components/schemas/Feeds"
        '400':
          description: Error

# -------------------------------------------------------------------------
# FOOD
# -------------------------------------------------------------------------
  /food:
    get:
      summary: Lists all the Food details till now 
      tags: [Food]
      responses:
        "200":
          description: The list of all food.
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Food'
    post:
      summary: Adds a new Food record.
      tags: [Food]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Food'
      responses: 
        '200':
          description: Created
        '400':
          description: Error  
  /food/{foodID}:
    get:
      summary: Returns information of a Food by ID.
      tags: [Food]
      parameters:
        - name: FoodID
          in: path
          required: true
          description: Autogenerated Food ID by MongoDB
          schema:
            type : string
      responses: 
        '200':
          description: OK
        '400':
          description: Record does not exist
        '500':
          description: Error

    delete:
      summary: Deletes a Food by ID.
      tags: [Food]
      parameters:
        - name: FoodID
          in: path
          required: true
          description: Autogenerated Food ID by MongoDB
          schema:
            type : string
      responses: 
        '200':
          description: Deleted
        '400':
          description: Record does not exist
        '500':
          description: Error
          
    patch:
      summary: Update a food information
      tags: [Food]
      parameters:
        - name: FoodID
          in: path
          required: true
          description: Autogenerated Food ID by MongoDB
          schema:
            type : string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Food'
      responses:
        '200':
          description: Feed Data Updated
          schema:
            $ref: "#/components/schemas/Food"
        '400':
          description: Error


# -------------------------------------------------------------------------
# ORAGANIZER
# -------------------------------------------------------------------------
...
components:
  schemas:
    Bookings:
      type: object
      required:
        - name
        - date
        - day
        - time
        - venues_name
        - silver
        - platinium
        - price
        - total_price
        - banner_image_url
        - grade

      properties:
        id:
          type: string
          description: The auto-generated id of the booking.
        name:
          type: string
          description: The name of the show.
        date:
          type: number
          description: Date of the show
        day:
          type: string
          description: Day of the show
        time:
          type: string
          description: Time of the show
        venues_name:
          type: string
          description: Name of the venue
        silver:
          type: array
          items:
            type: string
          description: Description of silver seats booked
        platinium:
          type: array
          items:
            type: string
          description: Description of platinium seats booked
        price:
          type: number
          description: Price of the seats
        total_price:
          type: number
          description: Total price 
        banner_image_url:
          type: string
          description: URL of the banner image
        grade:
          type: string
          description: Grade of the show  
    Events:
      type: object
      required:
        - name
        - location
        - feedback
        - is_popular
        - duration
        - about
        - genre
        - banner_image_url
        - languages
        - grade
        - rating
        - is_premier
        - release_date
        - cast

      properties:
        id:
          type: string
          description: The auto-generated id of the booking.
        name:
          type: string
          description: The name of the show.
        location:
          type: string
          description: Location of the show
        feedback:
          type: array
          items:
            type: string
          description: Feedback of the show
        is_popular:
          type: boolean
          items:
            type: string
          description: Show is popular or not
        duration:
          type: string
          description: Duration of the show
        about:
          type: string
          description: About of the show
        genre:
          type: array
          items:
            type: string
          description: Genre of the show
        banner_image_url:
          type: string
          description: URL of the banner image
        languages:
          type: string
          description: Language of the show 
        grade:
          type: string
          description: Grade of the show 
        rating:
          type: object
          description: Rating of the show 
        is_premier:
          type: boolean
          items:
            type: string
          description: Show is premier or not
        release_date:
          type: string
          description: Release Date of the show
        cast:
          type: array
          items:
            type: string
          description: Cast of the show
...
    Food:
      type: object
      required:
        - food_name
        - is_popcorn
        - is_coke
        - is_combo
        - food_tag
        - food_image
        - food_price
        - count
      properties:
        id:
          type: string
          description: The auto-generated id of the Food.
        food_name:
          type: string
          description: The name of the food.
        is_popcorn:
          type: boolean
          items:
            type: string
          description: Food is popcorn or not
        is_coke:
          type: boolean
          items:
            type: string
          description: Food is Coke or not
        is_combo:
          type: boolean
          items:
            type: string
          description: Food is combo pack or not
        food_tag:
          type: string
          description: The tag of the food.
        food_image:
          type: string
          description: URL of the food image
        food_price:
          type: string
          description: The price of the food
        count: 
          type: integer
          description: The count of the food
...